requirements.txt
matplotlib==3.5.2
numpy==1.21.6
opencv_python_headless==4.6.0.66
pandas==1.4.4
Pillow==9.3.0
pixellib==0.7.1
requests==2.28.1
scikit_image==0.19.2
scikit_learn==1.1.3
skimage==0.0
tensorflow==2.11.0
torch==1.12.1
tqdm==4.64.1
transformers==4.24.0